# locker-server/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.ErrorResponse:
    properties:
      error:
        example: invalid credentials
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      name:
        type: string
      phone_number:
        type: string
      student_id:
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  handlers.LogoutRequest:
    properties:
      access_token:
        description: '선택적: 요청 body에서 받기'
        type: string
      refresh_token:
        description: 반납할 refresh token
        type: string
    type: object
  handlers.LogoutResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.RefreshRequest:
    properties:
      access_token:
        description: '선택적: 블랙리스트용'
        type: string
      refresh_token:
        type: string
    type: object
  handlers.RefreshResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
info:
  contact: {}
  description: 사물함 선착순 예약 시스템의 백엔드 API 문서
  title: Locker Reservation API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 일치하는 사용자가 있으면 Access/Refresh 토큰 발급
      parameters:
      - description: 로그인 정보
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 로그인 (학번/이름/전화번호 확인)
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 현재 사용 중인 Access Token과 Refresh Token을 모두 무효화합니다.
      parameters:
      - default: Bearer
        description: Bearer {access_token}
        in: header
        name: Authorization
        type: string
      - description: 로그아웃 정보 (선택적)
        in: body
        name: payload
        schema:
          $ref: '#/definitions/handlers.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LogoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 로그아웃
      tags:
      - auth
  /auth/logout-all:
    post:
      consumes:
      - application/json
      description: 현재 사용자의 모든 Refresh Token을 무효화하여 모든 디바이스에서 로그아웃합니다.
      parameters:
      - default: Bearer
        description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LogoutResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 전체 로그아웃 (모든 디바이스)
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh 토큰을 사용하여 새로운 Access 토큰을 발급합니다.
      parameters:
      - description: 토큰 갱신 요청 정보
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RefreshResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 토큰 갱신
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Bearer {access_token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"